refactor: 상태이름 bookmark -> marked로 변경
